chore: 코드 분리
